менеджер предметов и подклассы предмет активируемый, предмет оружие, оружие ближнего и дальнего боя, оружие ближнего и дальнего боя на 5.56x45, а так же анимации атаки ближнего и дальнего боя для автомата

+++ Assets/Scripts/2.0/Items/Stats/itemsInterfaces.cs
@@ -3,8 +3,8 @@ using Unity.VisualScripting;
 
 public interface IMeleeWeapon
 {
-    float MinDamageMelee { get; }
-    float MaxDamageMelee { get; }
+    int MinDamageMelee { get; }
+    int MaxDamageMelee { get; }
     float CoolDownMelee { get; }
     float BaseStaminaMelee { get; }
     int AccuracyMelee { get; }
@@ -18,8 +18,8 @@ public interface IMeleeWeapon
 
 public interface IRangeWeapon
 {
-    float MinDamageRange { get; }
-    float MaxDamageRange { get; }
+    int MinDamageRange { get; }
+    int MaxDamageRange { get; }
     float CoolDownRange { get; }
     float BaseStaminaRange { get; }
     int AccuracyRange { get; }
